Release SimpleLLMFunc 0.8.1
