+++ SimpleLLMFunc/llm_decorator/llm_chat_decorator.py
@@ -1,6 +1,6 @@
+from __future__ import annotations
+
 import inspect
-import json
-from functools import wraps
 from typing import (
     Any,
     AsyncGenerator,
@@ -13,36 +13,47 @@ from typing import (
     TypeVar,
     Union,
     cast,
-    Literal,
 )
 
 from SimpleLLMFunc.base.react_loop import ReAct_loop
-from SimpleLLMFunc.llm_decorator.signature import (
-    parse_function_signature,
-    setup_log_context,
-)
-from SimpleLLMFunc.llm_decorator.invocation_builder import build_chat_invocation_spec
-from SimpleLLMFunc.runtime.selfref.session import SelfRefSession
-from SimpleLLMFunc.llm_decorator.prompt_contract import HISTORY_PARAM_NAMES
-from SimpleLLMFunc.llm_decorator.utils import (
-    process_tools,
-    remove_tool_best_practices_prompt_block,
-)
 from SimpleLLMFunc.base.types import CompileSource, DataFromAgentConfig
-from SimpleLLMFunc.interface.llm_interface import LLM_Interface
-from SimpleLLMFunc.runtime.selfref.state import (
-    MemoryHistory,
-    SELF_REFERENCE_KEY_OVERRIDE_TEMPLATE_PARAM,
-    SELF_REFERENCE_TOOLKIT_OVERRIDE_TEMPLATE_PARAM,
-    SelfReference,
-)
-from SimpleLLMFunc.runtime.selfref.state import _coerce_history_list
-from SimpleLLMFunc.runtime.selfref.context_ops import parse_data_from_selfref
-from SimpleLLMFunc.tool import Tool
-from SimpleLLMFunc.type import HistoryList, MessageList, NormalizedMessageParam
 from SimpleLLMFunc.hooks.events import ReactEndEvent
 from SimpleLLMFunc.hooks.stream import ReactOutput, ResponseYield, is_event_yield
-from SimpleLLMFunc.hooks.abort import AbortSignal, ABORT_SIGNAL_PARAM
+from SimpleLLMFunc.interface.llm_interface import LLM_Interface
+from SimpleLLMFunc.llm_decorator.chat_call_context import build_chat_call_context
+from SimpleLLMFunc.llm_decorator.chat_selfref import (
+    build_must_principles_prompt_block as _build_must_principles_prompt_block,
+    create_selfref_session,
+    extract_first_system_prompt_from_messages as _extract_first_system_prompt_from_messages,
+    extract_raw_history_reference as _extract_raw_history_reference,
+    finalize_self_reference_history as _finalize_self_reference_history,
+    react_end_event_has_fork_origin as _react_end_event_has_fork_origin,
+    remove_injected_prompt_blocks as _remove_injected_prompt_blocks,
+    remove_must_principles_prompt_block as _remove_must_principles_prompt_block,
+    remove_prompt_block as _remove_prompt_block,
+    remove_runtime_primitive_prompt_block as _remove_runtime_primitive_prompt_block,
+    resolve_runtime_self_reference_key,
+    resolve_self_reference_key as _resolve_self_reference_key,
+    seed_self_reference_system_prompt_if_missing as _seed_self_reference_system_prompt_if_missing,
+    set_history_argument as _set_history_argument,
+)
+from SimpleLLMFunc.llm_decorator.chat_toolkit import (
+    clone_toolkit_for_fork as _clone_toolkit_for_fork,
+    close_fork_cloned_pyrepls as _close_fork_cloned_pyrepls,
+    extract_self_reference_from_toolkit as _extract_self_reference_from_toolkit,
+    resolve_effective_self_reference as _resolve_effective_self_reference,
+    resolve_runtime_toolkit as _resolve_runtime_toolkit,
+)
+from SimpleLLMFunc.llm_decorator.chat_types import (
+    AGENT_FORK_TOOLKIT_FACTORY_ATTR,
+    AGENT_TEMPLATE_PARAMS_SUPPORT_ATTR,
+    DEFAULT_MAX_TOOL_CALLS,
+    ToolkitList,
+)
+from SimpleLLMFunc.llm_decorator.invocation_builder import build_chat_invocation_spec
+from SimpleLLMFunc.llm_decorator.prompt_contract import HISTORY_PARAM_NAMES
+from SimpleLLMFunc.llm_decorator.signature import setup_log_context
+from SimpleLLMFunc.llm_decorator.utils import process_tools
 from SimpleLLMFunc.observability.langfuse_client import (
     coerce_langfuse_metadata,
     get_langfuse_trace_context,
@@ -51,393 +62,31 @@ from SimpleLLMFunc.observability.langfuse_client import (
     update_langfuse_parent_span,
     update_langfuse_trace_name,
 )
+from SimpleLLMFunc.runtime.selfref.state import MemoryHistory, SelfReference
+from SimpleLLMFunc.tool import Tool
+from SimpleLLMFunc.type import HistoryList, MessageList, NormalizedMessageParam
 
-# Type aliases
-ToolkitList = List[Union[Tool, Callable[..., Awaitable[Any]]]]
-
-# Type variables
 T = TypeVar("T")
 P = ParamSpec("P")
 
-# Constants
-DEFAULT_MAX_TOOL_CALLS: Optional[int] = None  # No default tool-call cap
-_RUNTIME_PRIMITIVE_PROMPT_BLOCK_START = "<runtime_primitive_contract>"
-_RUNTIME_PRIMITIVE_PROMPT_BLOCK_END = "</runtime_primitive_contract>"
-_LEGACY_SELF_REFERENCE_PROMPT_BLOCK_START = "[SelfReference Memory Contract]"
-_LEGACY_SELF_REFERENCE_PROMPT_BLOCK_END = "[/SelfReference Memory Contract]"
-_MUST_PRINCIPLES_PROMPT_BLOCK_START = "<must_principles>"
-_MUST_PRINCIPLES_PROMPT_BLOCK_END = "</must_principles>"
-_AGENT_TEMPLATE_PARAMS_SUPPORT_ATTR = "__simplellmfunc_accepts_template_params__"
-_AGENT_FORK_TOOLKIT_FACTORY_ATTR = "__simplellmfunc_fork_toolkit_factory__"
-_FORK_CLONED_PYREPL_ATTR = "__simplellmfunc_fork_cloned_pyrepl__"
 
+def _build_chat_compile_source(invocation_spec: Any) -> CompileSource:
+    """Build the compatibility CompileSource payload from an InvocationSpec."""
 
-def _extract_raw_history_reference(
-    arguments: dict[str, Any],
-) -> Optional[MemoryHistory]:
-    """Extract the original history list object from bound call arguments."""
-
-    for history_param_name in HISTORY_PARAM_NAMES:
-        if history_param_name not in arguments:
-            continue
-
-        history = arguments[history_param_name]
-        if history is None:
-            return None
-
-        if isinstance(history, list) and all(
-            isinstance(item, dict) for item in history
-        ):
-            return cast(MemoryHistory, history)
-
-        return None
-
-    return None
-
-
-def _set_history_argument(
-    arguments: dict[str, Any],
-    history: MemoryHistory,
-) -> bool:
-    """Inject a history snapshot into bound arguments when history param exists."""
-
-    for history_param_name in HISTORY_PARAM_NAMES:
-        if history_param_name in arguments:
-            arguments[history_param_name] = list(history)
-            return True
-    return False
-
-
-def _clone_toolkit_for_fork(
-    base_toolkit: Optional[ToolkitList],
-    self_reference_override: Optional[SelfReference],
-) -> Optional[ToolkitList]:
-    if base_toolkit is None:
-        return None
-
-    from SimpleLLMFunc.builtin import PyRepl
-
-    cloned_toolkit: ToolkitList = []
-    repl_clones: Dict[int, PyRepl] = {}
-    backend_overrides: Optional[Dict[str, Any]] = None
-    if self_reference_override is not None:
-        backend_overrides = {
-            PyRepl.DEFAULT_SELF_REFERENCE_BACKEND_NAME: self_reference_override,
-        }
-
-    for item in base_toolkit:
-        if isinstance(item, Tool):
-            bound_instance = getattr(item.func, "__self__", None)
-            if isinstance(bound_instance, PyRepl):
-                original_repl_id = id(bound_instance)
-                if original_repl_id not in repl_clones:
-                    replacement_repl = bound_instance._clone_for_fork(
-                        backend_overrides=backend_overrides,
-                    )
-                    setattr(replacement_repl, _FORK_CLONED_PYREPL_ATTR, True)
-
-                    repl_clones[original_repl_id] = replacement_repl
-
-                replacement_repl = repl_clones[original_repl_id]
-                replacement_tool = next(
-                    tool for tool in replacement_repl.toolset if tool.name == item.name
-                )
-                cloned_toolkit.append(replacement_tool)
-                continue
-
-        cloned_toolkit.append(item)
-
-    return cloned_toolkit
-
-
-def _close_fork_cloned_pyrepls(toolkit: Optional[ToolkitList]) -> None:
-    if not toolkit:
-        return
-
-    from SimpleLLMFunc.builtin import PyRepl
-
-    closed_repl_ids: set[int] = set()
-
-    for item in toolkit:
-        if not isinstance(item, Tool):
-            continue
-
-        bound_instance = getattr(item.func, "__self__", None)
-        if not isinstance(bound_instance, PyRepl):
-            continue
-        if not bool(getattr(bound_instance, _FORK_CLONED_PYREPL_ATTR, False)):
-            continue
-
-        repl_id = id(bound_instance)
-        if repl_id in closed_repl_ids:
-            continue
-        closed_repl_ids.add(repl_id)
-
-        try:
-            bound_instance.close()
-        except Exception:
-            continue
-
-
-def _extract_self_reference_from_toolkit(
-    toolkit: Optional[ToolkitList],
-) -> Optional[SelfReference]:
-    if not toolkit:
-        return None
-
-    from SimpleLLMFunc.builtin import PyRepl
-
-    discovered: Dict[int, SelfReference] = {}
-
-    for item in toolkit:
-        if not isinstance(item, Tool):
-            continue
-
-        bound_instance = getattr(item.func, "__self__", None)
-        if not isinstance(bound_instance, PyRepl):
-            continue
-
-        default_backend = bound_instance.get_runtime_backend(
-            PyRepl.DEFAULT_SELF_REFERENCE_BACKEND_NAME
-        )
-        if isinstance(default_backend, SelfReference):
-            discovered[id(default_backend)] = default_backend
-            continue
-
-        for backend_name in bound_instance.list_runtime_backends():
-            backend_value = bound_instance.get_runtime_backend(backend_name)
-            if isinstance(backend_value, SelfReference):
-                discovered[id(backend_value)] = backend_value
-
-    if not discovered:
-        return None
-
-    return next(iter(discovered.values()))
-
-
-def _resolve_effective_self_reference(
-    explicit_self_reference: Optional[SelfReference],
-    toolkit: Optional[ToolkitList],
-) -> Optional[SelfReference]:
-    if explicit_self_reference is not None:
-        return explicit_self_reference
-    return _extract_self_reference_from_toolkit(toolkit)
-
-
-def _resolve_runtime_toolkit(
-    default_toolkit: Optional[ToolkitList],
-    template_params: Optional[Dict[str, Any]],
-) -> Optional[ToolkitList]:
-    if template_params is None:
-        return default_toolkit
-
-    override_toolkit = template_params.get(
-        SELF_REFERENCE_TOOLKIT_OVERRIDE_TEMPLATE_PARAM
-    )
-    if override_toolkit is None:
-        return default_toolkit
-
-    if not isinstance(override_toolkit, list):
-        raise ValueError("self_reference toolkit override must be a list")
-
-    if not default_toolkit:
-        return cast(Optional[ToolkitList], override_toolkit)
-
-    # Merge: start with default tools, then add override tools.
-    # When a tool with the same name exists in both, the override takes priority.
-    from SimpleLLMFunc.tool import Tool
-
-    seen_names: set[str] = set()
-    merged: ToolkitList = []
-
-    # Collect override tool names first (they take priority)
-    for item in override_toolkit:
-        if isinstance(item, Tool):
-            seen_names.add(item.name)
-        elif callable(item):
-            seen_names.add(getattr(item, "__name__", ""))
-
-    # Add default tools that are not overridden
-    for item in default_toolkit:
-        if isinstance(item, Tool):
-            if item.name not in seen_names:
-                merged.append(item)
-        elif callable(item):
-            name = getattr(item, "__name__", "")
-            if name not in seen_names:
-                merged.append(item)
-
-    # Append all override tools
-    merged.extend(override_toolkit)
-
-    return merged
-
-
-def _resolve_self_reference_key(
-    explicit_key: Optional[str],
-    func_name: str,
-) -> str:
-    if explicit_key is None:
-        return func_name
-
-    normalized = explicit_key.strip()
-    if not normalized:
-        raise ValueError("self_reference_key must be a non-empty string")
-    return normalized
-
-
-def _remove_prompt_block(system_prompt: str, start_marker: str, end_marker: str) -> str:
-    cleaned_prompt = system_prompt
-
-    while True:
-        start_index = cleaned_prompt.find(start_marker)
-        if start_index < 0:
-            break
-
-        end_index = cleaned_prompt.find(end_marker, start_index)
-        if end_index < 0:
-            cleaned_prompt = cleaned_prompt[:start_index]
-            break
-
-        cleaned_prompt = (
-            cleaned_prompt[:start_index] + cleaned_prompt[end_index + len(end_marker) :]
-        )
-
-    return cleaned_prompt
-
-
-def _remove_runtime_primitive_prompt_block(system_prompt: str) -> str:
-    cleaned_prompt = system_prompt
-    for start_marker, end_marker in (
-        (
-            _RUNTIME_PRIMITIVE_PROMPT_BLOCK_START,
-            _RUNTIME_PRIMITIVE_PROMPT_BLOCK_END,
+    return CompileSource(
+        data_from_agent_config=DataFromAgentConfig(
+            base_system_prompt=invocation_spec.prompt_contract.base_instruction,
+            template_params=(
+                dict(invocation_spec.template_params)
+                if invocation_spec.template_params is not None
+                else None
+            ),
+            tool_prompt_specs=list(invocation_spec.prompt_contract.tool_prompt_specs),
+            include_must_principles=invocation_spec.prompt_contract.include_must_principles,
         ),
-        (
-            _LEGACY_SELF_REFERENCE_PROMPT_BLOCK_START,
-            _LEGACY_SELF_REFERENCE_PROMPT_BLOCK_END,
-        ),
-    ):
-        cleaned_prompt = _remove_prompt_block(
-            cleaned_prompt,
-            start_marker,
-            end_marker,
-        )
-
-    return cleaned_prompt.strip()
-
-
-def _remove_injected_prompt_blocks(system_prompt: str) -> str:
-    cleaned_prompt = _remove_runtime_primitive_prompt_block(system_prompt)
-    cleaned_prompt = remove_tool_best_practices_prompt_block(cleaned_prompt)
-    cleaned_prompt = _remove_must_principles_prompt_block(cleaned_prompt)
-    return cleaned_prompt.strip()
-
-
-def _build_must_principles_prompt_block() -> str:
-    lines = [
-        _MUST_PRINCIPLES_PROMPT_BLOCK_START,
-        "<rule>Invoke tools through native structured tool_calls / function-calling fields.</rule>",
-        "<rule>Use assistant content for natural-language reasoning and final responses.</rule>",
-        "<rule>Keep tool invocation payloads in the native tool channel.</rule>",
-        _MUST_PRINCIPLES_PROMPT_BLOCK_END,
-    ]
-    return "\n".join(lines)
-
-
-def _remove_must_principles_prompt_block(system_prompt: str) -> str:
-    cleaned_prompt = _remove_prompt_block(
-        system_prompt,
-        _MUST_PRINCIPLES_PROMPT_BLOCK_START,
-        _MUST_PRINCIPLES_PROMPT_BLOCK_END,
+        data_from_selfref=invocation_spec.data_from_selfref,
+        input_messages=invocation_spec.transcript_seed.initial_messages,
     )
-    return cleaned_prompt.strip()
-
-
-def _extract_first_system_prompt_from_messages(messages: MessageList) -> Optional[str]:
-    for message in messages:
-        if message.get("role") != "system":
-            continue
-
-        content = message.get("content")
-        if isinstance(content, str):
-            return content
-
-    return None
-
-
-def _seed_self_reference_system_prompt_if_missing(
-    self_reference: SelfReference,
-    memory_key: str,
-    messages: MessageList,
-) -> None:
-    if self_reference.get_system_prompt(memory_key) is not None:
-        return
-
-    system_prompt = _extract_first_system_prompt_from_messages(messages)
-    if system_prompt is None:
-        return
-
-    # Keep memory store focused on durable prompt content. If the system prompt
-    # already contains auto-injected tool/runtime guidance blocks, store a
-    # cleaned version and let llm_chat inject guidance again per turn.
-    cleaned_system_prompt = _remove_injected_prompt_blocks(system_prompt)
-    system_prompt_for_store = cleaned_system_prompt or system_prompt
-
-    current_history: MemoryHistory = self_reference.snapshot_history(memory_key)
-    seeded_history: MemoryHistory = [
-        {"role": "system", "content": system_prompt_for_store},
-        *current_history,
-    ]
-    self_reference.replace_history(
-        key=memory_key,
-        messages=seeded_history,
-        strict=False,
-    )
-
-
-def _react_end_event_has_fork_origin(event: ReactEndEvent, origin: Any) -> bool:
-    origin_fork_id = getattr(origin, "fork_id", None)
-    if isinstance(origin_fork_id, str) and origin_fork_id:
-        return True
-
-    event_extra = getattr(event, "extra", None)
-    if not isinstance(event_extra, dict):
-        return False
-
-    raw_origin = event_extra.get("origin")
-    if not isinstance(raw_origin, dict):
-        return False
-
-    raw_fork_id = raw_origin.get("fork_id")
-    return isinstance(raw_fork_id, str) and bool(raw_fork_id)
-
-
-def _finalize_self_reference_history(
-    self_reference: SelfReference,
-    memory_key: str,
-    history: MemoryHistory,
-    *,
-    baseline_history_count: int,
-    base_system_prompt: str,
-) -> MemoryHistory:
-    resolved_history = cast(MemoryHistory, _coerce_history_list(history))
-    resolved_source = parse_data_from_selfref(resolved_history)
-    if (
-        resolved_source.experiences
-        or resolved_source.summary is not None
-        or resolved_source.summary_message is not None
-    ):
-        return self_reference.store_history(memory_key, resolved_history)
-
-    self_reference.consume_destructive_history_mutation(memory_key)
-    resolved_history = self_reference.merge_turn_history(
-        key=memory_key,
-        baseline_history_count=baseline_history_count,
-        updated_history=resolved_history,
-        commit=True,
-    )
-    return self_reference.store_history(memory_key, resolved_history)
 
 
 class LLMChat:
@@ -489,8 +138,8 @@ class LLMChat:
         self.__annotations__ = getattr(func, "__annotations__", {})
         self.__module__ = getattr(func, "__module__", __name__)
         self.__signature__ = self.signature_meta
-        setattr(self, _AGENT_TEMPLATE_PARAMS_SUPPORT_ATTR, True)
-        setattr(self, _AGENT_FORK_TOOLKIT_FACTORY_ATTR, self._build_fork_toolkit)
+        setattr(self, AGENT_TEMPLATE_PARAMS_SUPPORT_ATTR, True)
+        setattr(self, AGENT_FORK_TOOLKIT_FACTORY_ATTR, self._build_fork_toolkit)
 
         if self.self_reference is not None:
             self.self_reference.bind_agent_instance(
@@ -584,27 +233,17 @@ class LLMChat:
     async def __call__(
         self, *args: Any, **kwargs: Any
     ) -> AsyncGenerator[ReactOutput, None]:
-        abort_signal = kwargs.pop(ABORT_SIGNAL_PARAM, None)
-        if not isinstance(abort_signal, AbortSignal):
-            abort_signal = None
-        # Step 1: 解析函数签名
-        function_signature, template_params = parse_function_signature(
-            self.func,
-            args,
-            kwargs,
+        call_context = build_chat_call_context(
+            func=self.func,
+            args=args,
+            kwargs=kwargs,
+            default_toolkit=self.toolkit,
+            explicit_self_reference=self.self_reference,
         )
-        runtime_toolkit = _resolve_runtime_toolkit(self.toolkit, template_params)
-        effective_self_reference = _resolve_effective_self_reference(
-            self.self_reference,
-            runtime_toolkit,
-        )
-
-        # 构建用户任务提示（用于事件）
-        user_task_prompt = json.dumps(
-            function_signature.bound_args.arguments,
-            default=str,
-            ensure_ascii=False,
-        )
+        function_signature = call_context.signature
+        template_params = call_context.template_params
+        runtime_toolkit = call_context.runtime_toolkit
+        effective_self_reference = call_context.effective_self_reference
 
         toolkit_context_token = None
         active_memory_key_token = None
@@ -624,14 +263,12 @@ class LLMChat:
             )
 
         try:
-            # Step 2: 设置日志上下文
             async with setup_log_context(
                 func_name=function_signature.func_name,
                 trace_id=function_signature.trace_id,
                 arguments=function_signature.bound_args.arguments,
             ):
                 trace_context = get_langfuse_trace_context()
-                # 创建 Langfuse parent span
                 with langfuse_client.start_as_current_observation(
                     as_type="span",
                     name=f"{function_signature.func_name}_chat_call",
@@ -670,22 +307,11 @@ class LLMChat:
                                 previous_memory_key = (
                                     effective_self_reference._get_active_memory_key()
                                 )
-                                runtime_self_reference_key = self.self_reference_key
-                                if template_params is not None:
-                                    override_key = template_params.get(
-                                        SELF_REFERENCE_KEY_OVERRIDE_TEMPLATE_PARAM
-                                    )
-                                    if override_key is not None:
-                                        if not isinstance(override_key, str):
-                                            raise ValueError(
-                                                "self_reference key override must be a non-empty string"
-                                            )
-                                        runtime_self_reference_key = override_key
-
                                 resolved_self_reference_key = (
-                                    _resolve_self_reference_key(
-                                        runtime_self_reference_key,
-                                        function_signature.func_name,
+                                    resolve_runtime_self_reference_key(
+                                        explicit_key=self.self_reference_key,
+                                        func_name=function_signature.func_name,
+                                        template_params=template_params,
                                     )
                                 )
 
@@ -731,33 +357,19 @@ class LLMChat:
                                     )
                                 )
 
-                            # Step 3: 构建 invocation spec 与 selfref session
                             selfref_session = None
                             if (
                                 effective_self_reference is not None
                                 and resolved_self_reference_key is not None
                             ):
-                                selfref_session = SelfRefSession(
+                                selfref_session = create_selfref_session(
                                     backend=effective_self_reference,
                                     memory_key=resolved_self_reference_key,
                                     template_params=template_params,
                                     runtime_toolkit=runtime_toolkit,
                                     raw_history_reference=raw_history_reference,
                                     agent_instance=self,
-                                )
-                                selfref_session.history_authority = (
-                                    "external"
-                                    if raw_history_reference is not None
-                                    else (
-                                        "selfref"
-                                        if effective_self_reference.has_history(
-                                            resolved_self_reference_key
-                                        )
-                                        else "seed"
-                                    )
-                                )
-                                selfref_session.baseline_history_count = (
-                                    baseline_history_count
+                                    baseline_history_count=baseline_history_count,
                                 )
 
                             invocation_spec = build_chat_invocation_spec(
@@ -770,34 +382,16 @@ class LLMChat:
                                 raw_history_reference=raw_history_reference,
                             )
                             messages = invocation_spec.transcript_seed.initial_messages
-                            # Compatibility payload for existing step tests / patched runners.
-                            # The formal LLM conversion still uses invocation_spec -> compile_pipeline.
-                            compile_source = CompileSource(
-                                data_from_agent_config=DataFromAgentConfig(
-                                    base_system_prompt=invocation_spec.prompt_contract.base_instruction,
-                                    template_params=(
-                                        dict(invocation_spec.template_params)
-                                        if invocation_spec.template_params is not None
-                                        else None
-                                    ),
-                                    tool_prompt_specs=list(
-                                        invocation_spec.prompt_contract.tool_prompt_specs
-                                    ),
-                                    include_must_principles=invocation_spec.prompt_contract.include_must_principles,
-                                ),
-                                data_from_selfref=invocation_spec.data_from_selfref,
-                                input_messages=invocation_spec.transcript_seed.initial_messages,
-                            )
+                            compile_source = _build_chat_compile_source(invocation_spec)
 
                             if selfref_session is not None:
                                 _seed_self_reference_system_prompt_if_missing(
                                     effective_self_reference,
-                                    resolved_self_reference_key,
+                                    cast(str, resolved_self_reference_key),
                                     messages,
                                 )
                                 react_hooks = selfref_session
 
-                            # Step 4: 执行 ReAct 循环（流式）
                             tool_param, tool_map = process_tools(
                                 runtime_toolkit,
                                 function_signature.func_name,
@@ -811,8 +405,8 @@ class LLMChat:
                                 max_tool_calls=self.max_tool_calls,
                                 stream=self.stream,
                                 trace_id=function_signature.trace_id,
-                                user_task_prompt=user_task_prompt,
-                                abort_signal=abort_signal,
+                                user_task_prompt=call_context.user_task_prompt,
+                                abort_signal=call_context.abort_signal,
                                 hooks=react_hooks,
                                 invocation_spec=invocation_spec,
                                 toolkit=runtime_toolkit,
@@ -843,11 +437,7 @@ class LLMChat:
                                             resolved_self_reference_key,
                                             cast(MemoryHistory, history),
                                             baseline_history_count=baseline_history_count,
-                                            base_system_prompt=(
-                                                compile_source.data_from_agent_config.base_system_prompt
-                                                if compile_source is not None
-                                                else ""
-                                            ),
+                                            base_system_prompt=compile_source.data_from_agent_config.base_system_prompt,
                                         )
                                         output.messages = cast(
                                             MessageList, active_history
@@ -883,11 +473,7 @@ class LLMChat:
                                             MemoryHistory, output.event.final_messages
                                         ),
                                         baseline_history_count=baseline_history_count,
-                                        base_system_prompt=(
-                                            compile_source.data_from_agent_config.base_system_prompt
-                                            if compile_source is not None
-                                            else ""
-                                        ),
+                                        base_system_prompt=compile_source.data_from_agent_config.base_system_prompt,
                                     )
                                     output.event.final_messages = cast(
                                         HistoryList,
@@ -916,7 +502,6 @@ class LLMChat:
                                 },
                             )
                         except Exception as exc:
-                            # 更新 span 错误信息
                             chat_span.update(
                                 output={"error": str(exc)},
                             )
@@ -1009,3 +594,28 @@ def llm_chat(
 
 
 async_llm_chat = llm_chat
+
+__all__ = [
+    "DEFAULT_MAX_TOOL_CALLS",
+    "LLMChat",
+    "ToolkitList",
+    "async_llm_chat",
+    "llm_chat",
+    "_build_must_principles_prompt_block",
+    "_clone_toolkit_for_fork",
+    "_close_fork_cloned_pyrepls",
+    "_extract_first_system_prompt_from_messages",
+    "_extract_raw_history_reference",
+    "_extract_self_reference_from_toolkit",
+    "_finalize_self_reference_history",
+    "_react_end_event_has_fork_origin",
+    "_remove_injected_prompt_blocks",
+    "_remove_must_principles_prompt_block",
+    "_remove_prompt_block",
+    "_remove_runtime_primitive_prompt_block",
+    "_resolve_effective_self_reference",
+    "_resolve_runtime_toolkit",
+    "_resolve_self_reference_key",
+    "_seed_self_reference_system_prompt_if_missing",
+    "_set_history_argument",
+]
