+++ skills/simplellmfunc-developer/reference/project-map.md
@@ -19,11 +19,16 @@ SimpleLLMFunc/
 Primary user entrypoints:
 
 - `llm_function_decorator.py`
-- `llm_chat_decorator.py`
-- `selfref_sync.py` (legacy compatibility shim; active SelfRef session binding is in `LLMChat` + `runtime/selfref/session.py`) (legacy compatibility shim; active SelfRef session binding is in `LLMChat` + `runtime/selfref/session.py`)
-- `steps/common/`
-- `steps/function/`
-- `steps/chat/`
+- `llm_chat_decorator.py` (public facade / `LLMChat` callable)
+- `chat_call_context.py`
+- `chat_selfref.py`
+- `chat_toolkit.py`
+- `chat_types.py`
+- `selfref_sync.py` (legacy compatibility shim; active SelfRef session binding is in `LLMChat` + `runtime/selfref/session.py`)
+- `invocation_spec.py`
+- `invocation_builder.py`
+- `prompt_contract.py`
+- `signature.py`
 - `utils/tools.py`
 
 This layer turns typed Python call signatures into prompt-building, ReAct execution, response parsing, and lifecycle-aware integrations such as `llm_chat` <-> `SelfReference` synchronization.
@@ -53,16 +58,31 @@ This layer powers the `runtime.*` namespace used inside `PyRepl`.
 
 Key current selfref split:
 
+- `runtime/selfref/state.py`: `SelfReference` public facade, lifecycle, compatibility exports
+- `runtime/selfref/store.py`: durable history/source store
+- `runtime/selfref/active_turn.py`: active key/fork/toolkit/template contextvars and active ReAct state lookup
+- `runtime/selfref/mutations.py`: pending compaction/context/destructive mutation queues
+- `runtime/selfref/memory_api.py`: `self_reference.memory[...]` handle/proxy API
+- `runtime/selfref/context_memory.py`: context snapshots, experience CRUD, compaction commit, direct memory editing
+- `runtime/selfref/agent_binding.py`: recursive agent callable binding
+- `runtime/selfref/fork_manager.py`: fork/spawn/gather lifecycle and result materialization
+- `runtime/selfref/fork_utils.py`: fork helpers/constants
 - `runtime/selfref/context_ops.py`: pure parse/render/canonicalize helpers for context messages
-- `runtime/selfref/state.py`: persistent memory state, mutation, validation, compaction queueing, and fork-aware behavior
 
 ### `SimpleLLMFunc/builtin/`
 
 End-user built-ins:
 
-- `pyrepl.py`
+- `pyrepl.py` (public facade)
+- `pyrepl_execution.py`
+- `pyrepl_worker_client.py`
+- `pyrepl_worker_mixin.py`
+- `pyrepl_primitive_host.py`
+- `pyrepl_tools.py`
+- `pyrepl_audit.py`
+- `pyrepl_input_bridge.py`
+- `pyrepl_input_mixin.py`
 - `file_tools.py`
-- `self_reference.py`
 
 These builtins expose high-level behavior on top of lower-level runtime and tool plumbing.
 
@@ -101,10 +121,11 @@ Support layers such as:
 - Decorator behavior bug: start in `llm_decorator/`, then trace into `base/`.
 - `llm_chat` + selfref sync issue: start in `llm_chat_decorator.py` (`LLMChat`) and `runtime/selfref/session.py`, then inspect `runtime/selfref/state.py` and the relevant decorator entrypoint.
 - Tool schema or prompt-injection issue: start in `tool/tool.py` and `llm_decorator/utils/tools.py`.
-- Runtime primitive issue: start in `runtime/primitives.py`, then inspect `builtin/pyrepl.py` or the relevant builtin backend.
+- Runtime primitive issue: start in `runtime/primitives.py`, then inspect `builtin/pyrepl_primitive_host.py` or the relevant builtin backend.
 - Selfref context parse/render issue: start in `runtime/selfref/context_ops.py` before changing stateful code.
-- Event-stream issue: start in `hooks/`, then inspect decorator step modules.
-- ReAct termination / hook-order issue: start in `base/ReAct.py` and follow the phase helpers plus `before_finalize` call sites.
+- Selfref context memory or compaction issue: start in `runtime/selfref/context_memory.py`, then inspect `runtime/selfref/session.py`.
+- Selfref fork lifecycle issue: start in `runtime/selfref/fork_manager.py` and `runtime/selfref/fork_utils.py`.
+- Event-stream issue: start in `hooks/`, then inspect `base/react_loop.py` and decorator facade modules.
+- ReAct termination / hook-order issue: start in `base/react_loop.py` plus `base/react_hooks.py` and `before_finalize` call sites.
 - Provider transport issue: start in the relevant adapter under `interface/` (`openai_compatible.py` or `openai_responses_compatible.py`) plus related tests.
-- Selfref fork context-construction issue: start in `runtime/selfref/state.py`; check how pre-fork history is stripped/materialized before touching decorator wrappers.
 - User-facing docs mismatch: source code and tests win; then patch `mintlify_docs/` and maybe `README.md`.
